+++ src/components/ClientGrid.js
@@ -1,5 +1,5 @@
 import React from 'react'
 
-const ClientGrid = () => <p>Clients Ive worked with in the past.</p>
+const ClientGrid = () => <p className={`title h6 text-white`}>Clients Ive worked with in the past.</p>
 
 export default ClientGrid