+++ src/components/header.js
@@ -4,7 +4,7 @@ import UnderlostSVG from './svg/underlost'
 const Header = () => (
   <div className={`row no-gutters`}>
     <div className={`col-md-5 offset-md-7 pt-3`}>
-      <header className={`site-header site-header-sm mt-5`}>
+      <header className={`site-header site-header-sm`}>
         <UnderlostSVG />
       </header>
     </div>