+++ src/components/portfolioItem.js
@@ -7,20 +7,20 @@ import Image from './Image'
 
 
 const PortfolioItem = props => (
-  <Fade big cascade>
-    <div key={props.guid} className={`project-item portfolio-item`}>
+  <div key={props.guid} className={`project-item portfolio-item col-6`}>
+    <Fade big>
       <div className={`project-cover`}>
         <Image filename={props.cover} alt={props.title} />
       </div>
       <div className={`mt-3`}>
         <span className={`h6 title guid`}>{props.guid}</span>
-        <h3 className={`h4 headline text-uppercase`}>
+        <h3 className={`h4 headline text-uppercase mb-0`}>
           <Link to={props.permalink}>{props.title}</Link>
         </h3>
       </div>
-      <div className={`lead`}>{props.description}</div>
-    </div>
-  </Fade>
+      <p className={`sr-only`}>{props.description}</p>
+    </Fade>
+  </div>
 )
 
 PortfolioItem.propTypes = {