+++ CHANGELOG.md
@@ -1,24 +1,39 @@
-###6.0.0 (05/14/2019)
+# CHANGELOG
+
+---
+
+## 6.1.0 (02/11/2020)
+
+* Still WIP
+* Layout designed
+* Changed formatting for portfolio page to include case studies
+
+## 6.0.0 (05/14/2019)
+
 * WIP
 * Switched to React.JS and Gatsby
 
-###5.1.0 (04/28/2018)
+## 5.1.0 (04/28/2018)
+
 * Updated bootstrap to 4.1
 * New design
 * underlost is probably not an Undertale game.
 
-###5.0.0 (02/16/2018)
+## 5.0.0 (02/16/2018)
+
 * Updated bootstrap
 * Updated Font-Awesome
 * New design
 
-###4.3.0 (01/14/2018)
+## 4.3.0 (01/14/2018)
+
 * Updated footer
 * Updated bootstrap
 * Changed the way pages are laid out in Jekyll.
 * Updated portfolio.
 
-### 4.2.0 (05/23/2017)
+## 4.2.0 (05/23/2017)
+
 * Started a Changelog
 * Removed Bower, moved everything to Yarn/package.json
 * Updated portfolio navigation